requirements.txt
pymata4
pyautogui
python-vlc